+++ src/main/java/org/guzzing/studayserver/domain/auth/jwt/logout/LogoutAuthenticationFilter.java
@@ -1,43 +1,41 @@
-package org.guzzing.studayserver.domain.auth.jwt;
+package org.guzzing.studayserver.domain.auth.jwt.logout;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.guzzing.studayserver.domain.auth.jwt.AuthToken;
+import org.guzzing.studayserver.domain.auth.jwt.AuthTokenProvider;
+import org.guzzing.studayserver.domain.auth.jwt.JwtHeaderUtil;
+import org.guzzing.studayserver.domain.auth.service.AuthService;
+import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
-import org.guzzing.studayserver.domain.auth.service.AuthService;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
+public class LogoutAuthenticationFilter extends OncePerRequestFilter {
 
     private final static String HEADER_AUTHORIZATION = "Authorization";
     private final static String TOKEN_PREFIX = "Bearer ";
 
     private final AuthTokenProvider tokenProvider;
+    private final AuthService authService;
 
-    public JwtAuthenticationFilter(AuthTokenProvider tokenProvider) {
+    public LogoutAuthenticationFilter(AuthTokenProvider tokenProvider, AuthService authService) {
         this.tokenProvider = tokenProvider;
+        this.authService = authService;
     }
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                    HttpServletResponse response,
-                                    FilterChain filterChain) throws ServletException, IOException {
+    protected void doFilterInternal(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            FilterChain filterChain) throws ServletException, IOException {
 
         final String AUTHORIZATION_HEADER = request.getHeader(HEADER_AUTHORIZATION);
 
         if (AUTHORIZATION_HEADER != null && AUTHORIZATION_HEADER.startsWith(TOKEN_PREFIX)) {
             String tokenStr = JwtHeaderUtil.getAccessToken(request);
-            AuthToken token = tokenProvider.convertAuthToken(tokenStr);
-
-            if (token.isValidTokenClaims()) {
-                Authentication authentication = tokenProvider.getAuthentication(token);
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
+            authService.isLogout(tokenStr);
 
             filterChain.doFilter(request, response);
             return;
